Exploratory Refactoring #2: all tests passing
