+++ amf-shapes/shared/src/main/scala/amf/plugins/document/webapi/parser/spec/declaration/emitters/ShapeEmitterContext.scala
@@ -1,6 +1,7 @@
 package amf.plugins.document.webapi.parser.spec.declaration.emitters
 
 import amf.client.remod.amfcore.config.ShapeRenderOptions
+import amf.core.emitter.BaseEmitters.MapEntryEmitter
 import amf.core.emitter.{Emitter, EntryEmitter, PartEmitter, SpecOrdering}
 import amf.core.errorhandling.ErrorHandler
 import amf.core.metamodel.Field
@@ -9,11 +10,15 @@ import amf.core.model.domain.extensions.{DomainExtension, ShapeExtension}
 import amf.core.model.domain.{DomainElement, Linkable, RecursiveShape, Shape}
 import amf.core.parser.FieldEntry
 import amf.core.remote.Vendor
+import amf.plugins.document.webapi.parser.CommonOasTypeDefMatcher
+import amf.plugins.document.webapi.parser.spec.declaration.common.ExternalLinkQuery.queryResidenceUnitOfLinked
 import amf.plugins.document.webapi.parser.spec.declaration.emitters.annotations.FacetsInstanceEmitter
 import amf.plugins.document.webapi.parser.spec.declaration.{CustomFacetsEmitter, SchemaVersion}
 import amf.plugins.document.webapi.parser.spec.oas.emitters.OasLikeExampleEmitters
-import amf.plugins.domain.shapes.models.{Example, TypeDef}
-import org.yaml.model.{YDocument, YNode}
+import amf.plugins.domain.shapes.metamodel.NodeShapeModel
+import amf.plugins.domain.shapes.models.{Example, TypeDef, UnionShape}
+import amf.plugins.domain.webapi.annotations.TypePropertyLexicalInfo
+import org.yaml.model.{YDocument, YNode, YType}
 
 trait SpecAwareEmitterContext {
   def factoryIsOas3: Boolean
@@ -25,11 +30,25 @@ trait SpecAwareEmitterContext {
 
 trait ShapeEmitterContext extends SpecAwareEmitterContext {
 
-  def externalLink(shape: Shape, references: Seq[BaseUnit]): Option[BaseUnit]
+  def externalLink(link: Linkable, refs: Seq[BaseUnit]): Option[BaseUnit] = queryResidenceUnitOfLinked(link, refs)
 
   def toOasNext: ShapeEmitterContext
 
-  def ramlTypePropertyEmitter(value: String, shape: Shape): Option[EntryEmitter]
+  def ramlTypePropertyEmitter(typeName: String, shape: Shape): Option[MapEntryEmitter] = {
+    shape.fields.?(NodeShapeModel.Inherits) match {
+      case None =>
+        // If the type is union and anyOf is empty it isn't resolved and type will be emitted in the UnionEmitter
+        if (typeName == "union" && shape.asInstanceOf[UnionShape].anyOf.nonEmpty) None
+        else {
+          shape.annotations.find(classOf[TypePropertyLexicalInfo]) match {
+            case Some(lexicalInfo) =>
+              Some(MapEntryEmitter("type", typeName, YType.Str, lexicalInfo.range.start))
+            case _ => None
+          }
+        }
+      case _ => None
+    }
+  }
 
   def localReference(shape: Shape): PartEmitter
 
@@ -39,15 +58,22 @@ trait ShapeEmitterContext extends SpecAwareEmitterContext {
                             ordering: SpecOrdering,
                             schemaPath: Seq[(String, String)]): Emitter
 
-  def oasMatchType(get: TypeDef): String
+  def oasMatchType(get: TypeDef): String = CommonOasTypeDefMatcher.matchType(get)
 
-  def oasMatchFormat(typeDef: TypeDef): Option[String]
+  def oasMatchFormat(typeDef: TypeDef): Option[String] = CommonOasTypeDefMatcher.matchFormat(typeDef)
 
   def schemasDeclarationsPath: String
 
   def arrayEmitter(asOasExtension: String, f: FieldEntry, ordering: SpecOrdering): EntryEmitter
 
-  def oasTypePropertyEmitter(str: String, shape: Shape): EntryEmitter
+  def oasTypePropertyEmitter(typeName: String, shape: Shape): EntryEmitter = {
+    shape.annotations.find(classOf[TypePropertyLexicalInfo]) match {
+      case Some(lexicalInfo) =>
+        MapEntryEmitter("type", typeName, YType.Str, lexicalInfo.range.start)
+      case None =>
+        MapEntryEmitter("type", typeName)
+    }
+  }
 
   def customFacetsEmitter(f: FieldEntry, ordering: SpecOrdering, references: Seq[BaseUnit]): CustomFacetsEmitter
