+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/contexts/SpecEmitterContext.scala
@@ -9,6 +9,7 @@ import amf.core.model.domain.extensions.{CustomDomainProperty, DomainExtension, 
 import amf.core.model.domain.{DomainElement, Linkable, Shape}
 import amf.core.parser.FieldEntry
 import amf.core.remote._
+import amf.plugins.document.webapi.parser.spec.declaration.common.ExternalLinkQuery.queryResidenceUnitOfLinked
 import amf.plugins.document.webapi.parser.spec.declaration._
 import amf.plugins.document.webapi.parser.spec.declaration.emitters.annotations.{
   AnnotationEmitter,
@@ -39,31 +40,6 @@ abstract class SpecEmitterContext(val eh: ErrorHandler, refEmitter: RefEmitter, 
   def arrayEmitter(key: String, f: FieldEntry, ordering: SpecOrdering, valuesTag: YType = YType.Str): EntryEmitter =
     ArrayEmitter(key, f, ordering, forceMultiple = false, valuesTag)
 
-  def oasTypePropertyEmitter(typeName: String, shape: Shape): MapEntryEmitter = {
-    shape.annotations.find(classOf[TypePropertyLexicalInfo]) match {
-      case Some(lexicalInfo) =>
-        MapEntryEmitter("type", typeName, YType.Str, lexicalInfo.range.start)
-      case None =>
-        MapEntryEmitter("type", typeName)
-    }
-  }
-
-  def ramlTypePropertyEmitter(typeName: String, shape: Shape): Option[MapEntryEmitter] = {
-    shape.fields.?(NodeShapeModel.Inherits) match {
-      case None =>
-        // If the type is union and anyOf is empty it isn't resolved and type will be emitted in the UnionEmitter
-        if (typeName == "union" && shape.asInstanceOf[UnionShape].anyOf.nonEmpty) None
-        else {
-          shape.annotations.find(classOf[TypePropertyLexicalInfo]) match {
-            case Some(lexicalInfo) =>
-              Some(MapEntryEmitter("type", typeName, YType.Str, lexicalInfo.range.start))
-            case _ => None
-          }
-        }
-      case _ => None
-    }
-  }
-
   private var emittingDeclarations: Boolean = false
 
   def runAsDeclarations(fn: () => Unit): Unit = {
@@ -75,22 +51,6 @@ abstract class SpecEmitterContext(val eh: ErrorHandler, refEmitter: RefEmitter, 
   def filterLocal[T <: DomainElement](elements: Seq[T]): Seq[T] = {
     if (!emittingDeclarations) elements
     else elements.filter(!_.fromLocal())
-  }
-
-  def externalLink(link: Linkable, refs: Seq[BaseUnit]): Option[BaseUnit] = {
-    link.linkTarget match {
-      case Some(element) =>
-        val linkTarget = element.id
-        refs.find {
-          case fragment: EncodesModel =>
-            Option(fragment.encodes).exists(_.id == linkTarget)
-          case library: DeclaresModel =>
-            library.declares.exists(_.id == linkTarget)
-          case _ => false
-        }
-
-      case _ => None
-    }
   }
 }
